Revert "refactor: colocate cart page components with route"
This reverts commit a6fd7d6b7d8c91cb41e5ee34848d8893cbb17e56.
